pound to kg 변환 시 소수점 문제로 데이터 변경 및 변환 계산 변경

+++ src/main/java/com/example/healthcare/application/exercise/controller/dto/CreateUserExerciseSetDTO.java
@@ -9,7 +9,7 @@ import jakarta.validation.constraints.NotNull;
 public record CreateUserExerciseSetDTO(
   @NotNull @Min(1) Long setNumber,
   @NotNull ExerciseSetType exerciseSetType,
-  Long weight,
+  Double weight,
   Integer reps,
   Long time,
   @NotNull Boolean complete
@@ -32,7 +32,7 @@ public record CreateUserExerciseSetDTO(
   }
 
   @Override
-  public Long getWeight() {
+  public Double getWeight() {
     return weight();
   }
